% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CO2}
\alias{CO2}
\title{Carbon emissions around the world (2014)}
\format{A tibble data frame with 116 rows and 3 variables.}
\usage{
CO2
}
\description{
Carbon emissions around the world (2014)
}
\keyword{datasets}
